requirements.txt
httpx
pydantic
pydantic-settings
python-dotenv
loguru
pytest
pytest-mock
pytest-asyncio 